Use single span to display emoji.
The outer highlight can be done with box-shadow + border-radius,
but there is one more problem: the emojis are now to close to
each other.
In the original version, the spacing was added by padding in the
inner span. With single span, padding doesn't work and if we use
margin then we have "gaps" between emojis, so it is harder to
select with the mouse.
And "border" also doesn't work as we use "display: inline-block".