Fix $refs usage for Vue 3.
Accodring to the [migration guide](https://v3.vuejs.org/guide/migration/array-refs.html),
the automatic array for $refs is not supported anymore and there is a new feature
to bind a function as "ref" which we also can not use to keep Vue 2 support.

So instead I am using a dynamic names for category refs (category_0, category_1, etc).