Add accessibility properties to emoji picker.
It now works with VoiceOver, it is possible to loop through the
picker elements and "click" them.

It does not work with regular keyboard control (no Voice Over, just
arrow keys).

This is actually expected, and the parent project [change](https://github.com/missive/emoji-mart/pull/284)
also states that:

> should I now be able to type, then navigate results via arrow keys?

Yes, in terms of improving the accessibility of emoji-mart, I didn't make it a full autocomplete pattern
where you can press the arrow keys to navigate between results.
You have to just use the regions to move around after typing into the search bar.

Certainly not a perfect design a11y-wise, but it's at least usable with a keyboard, whereas before it wasn't.