fix: declare all emits in all components
This is the same as https://github.com/serebrov/emoji-mart-vue/pull/243
but in the other places that were not covered there.

We encountered this in Safari 17.4.1 where the categories would change
on Vue click event correctly but right after that the native click with
ClickEvent object as it's first argument would be processed and make the
picker completely empty with no way back.